@@ -2,10 +2,6 @@
   "name": "angular-interceptors",
   "version": "2.1.0",
   "license": "GPL-3.0",
-  "engines": {
-    "node": ">=8.9.0",
-    "yarn": ">=1.3.2"
-  },
   "scripts": {
     "package": "ng-packagr -p package.json",
     "format": "clang-format --glob=\"@(src)/**/*.@(ts|js)\" -i && clang-format --glob=\"*.@(ts|js)\" -i",
@@ -14,25 +10,33 @@
     "postrelease": "git push --follow-tags origin master && yarn package && conventional-github-releaser -p angular"
   },
   "devDependencies": {
-    "@angular/common": ">=5.2.3 <6.0.0",
-    "@angular/compiler": ">=5.2.3 <6.0.0",
-    "@angular/compiler-cli": ">=5.2.3 <6.0.0",
-    "@angular/core": ">=5.2.3 <6.0.0",
-    "@angular/language-service": ">=5.2.3 <6.0.0",
-    "@angular/platform-browser": ">=5.2.3 <6.0.0",
-    "@angular/platform-browser-dynamic": ">=5.2.3 <6.0.0",
+    "@angular/cli": "^1.7.3",
+    "@angular/common": ">=6.0.0 <7.0.0",
+    "@angular/compiler": ">=6.0.0 <7.0.0",
+    "@angular/compiler-cli": ">=6.0.0 <7.0.0",
+    "@angular/core": ">=6.0.0 <7.0.0",
+    "@angular/language-service": ">=6.0.0 <7.0.0",
+    "@angular/platform-browser": ">=6.0.0 <7.0.0",
+    "@angular/platform-browser-dynamic": ">=6.0.0 <7.0.0",
     "clang-format": "^1.1.0",
-    "codelyzer": "^4.1.0",
-    "conventional-github-releaser": "^2.0.0",
-    "core-js": "2.x",
-    "ng-packagr": "1.7.0",
-    "rxjs": "5.x",
+    "codelyzer": "^4.3.0",
+    "conventional-github-releaser": "^2.0.2",
+    "core-js": "^2.5.6",
+    "jasmine": "^3.1.0",
+    "jasmine-core": "^3.1.0",
+    "karma": "^2.0.0",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-coverage-istanbul-reporter": "^1.4.2",
+    "karma-jasmine": "^1.1.1",
+    "karma-jasmine-html-reporter": "^1.0.0",
+    "ng-packagr": "^3.0.0-rc.3",
+    "rxjs": "^6.1.0",
     "standard-version": "^4.3.0",
-    "tsickle": "^0.26",
+    "tsickle": "^0.28.0",
     "tslib": "^1.9.0",
-    "tslint": "^5.9.1",
-    "typescript": ">=2.4.2 <2.7",
-    "zone.js": "0.x"
+    "tslint": "^5.10.0",
+    "typescript": "2.7.x",
+    "zone.js": "^0.8.x"
   },
   "ngPackage": {
     "dest": "lib",
