@@ -1,3 +1,3 @@
 export {CacheInterceptorModule, MAX_AGE_MS} from './cache.module';
 export {EnsureHttpsInterceptorModule} from './ensure-https.module';
-export {PrefixUrlInterceptorModule, PREFIX} from './prefix-url.module';
+export {PREFIX, PrefixUrlInterceptorModule} from './prefix-url.module';
